requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
aziraclient==0.1.5
bidict==0.22.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
fastapi==0.104.1
fastapi-socketio==0.0.10
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
iso8601==2.1.0
jwt==1.3.1
kombu==5.3.3
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyTelegramBotAPI==4.14.0
pytest==7.4.3
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pyzmq==25.1.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
vine==5.1.0
websockets==12.0
wsproto==1.2.0
zmq==0.0.0
